feat: add update curriculum vitae server action
